+++ deode/argparse_wrapper.py
@@ -8,6 +8,8 @@ from . import GeneralConstants
 from .commands_functions import (
     create_exp,
     doc_config,
+    namelist_convert,
+    namelist_format,
     namelist_integrate,
     run_task,
     show_config,
@@ -17,6 +19,7 @@ from .commands_functions import (
     start_suite,
 )
 from .config_parser import ConfigParserDefaults
+from .namelist import NamelistConverter
 
 
 def get_parsed_args(program_name=GeneralConstants.PACKAGE_NAME, argv=None):
@@ -285,10 +288,12 @@ def get_parsed_args(program_name=GeneralConstants.PACKAGE_NAME, argv=None):
 
     # namelist subparser
     parser_namelist = subparsers.add_parser(
-        "namelist", help="Namelist show (output) or integrate (input)"
+        "namelist",
+        help="Namelist show (output), integrate (input), "
+        + "convert (input, output), format (input, output)",
     )
     namelist_command_subparsers = parser_namelist.add_subparsers(
-        title="integrate",
+        title="namelist",
         dest="namelist_what",
         required=True,
         description=(
@@ -338,5 +343,78 @@ def get_parsed_args(program_name=GeneralConstants.PACKAGE_NAME, argv=None):
         default=None,
     )
     parser_namelist_integrate.set_defaults(run_command=namelist_integrate)
+
+    # namelist convert
+    parser_namelist_convert = namelist_command_subparsers.add_parser(
+        "convert",
+        help="Convert a namelist (ftn or yml) to a new Cycle",
+        parents=[common_parser],
+    )
+    parser_namelist_convert.add_argument(
+        "-n",
+        "--namelist",
+        type=str,
+        help="Input namelist definition filename",
+        required=True,
+        default=None,
+    )
+    parser_namelist_convert.add_argument(
+        "-o",
+        "--output",
+        type=str,
+        help="Output namelist definition filename",
+        required=True,
+        default=None,
+    )
+    parser_namelist_convert.add_argument(
+        "--from-cycle",
+        type=str,
+        help="Cycle of input namelist",
+        choices=NamelistConverter.get_known_cycles(),
+        required=True,
+        default=None,
+    )
+
+    parser_namelist_convert.add_argument(
+        "--to-cycle",
+        type=str,
+        help="Cycle of output namelist",
+        choices=NamelistConverter.get_known_cycles(),
+        required=True,
+        default=None,
+    )
+
+    parser_namelist_convert.add_argument(
+        "--format", "-fmt", help="Input format", choices=["yaml", "ftn"], default="yaml"
+    )
+    parser_namelist_convert.set_defaults(run_command=namelist_convert)
+
+    # namelist format
+    parser_namelist_format = namelist_command_subparsers.add_parser(
+        "format",
+        help="Format a namelist (ftn or yml) ",
+        parents=[common_parser],
+    )
+    parser_namelist_format.add_argument(
+        "-n",
+        "--namelist",
+        type=str,
+        help="Input namelist definition filename",
+        required=True,
+        default=None,
+    )
+    parser_namelist_format.add_argument(
+        "-o",
+        "--output",
+        type=str,
+        help="Output namelist definition filename",
+        required=True,
+        default=None,
+    )
+
+    parser_namelist_format.add_argument(
+        "--format", "-fmt", help="Input format", choices=["yaml", "ftn"], default="yaml"
+    )
+    parser_namelist_format.set_defaults(run_command=namelist_format)
 
     return main_parser.parse_args(argv)